TDLM

Documentation for TDLM.

TDLM.Simulate.Noise

TDLM.Simulate.rand_cov

Simulate

TDLM.Simulate.Noise — Type

Noise()

Add noise of a certain distribution to an array adopting the shape an structure of the array. If nothing else specified, randn is used, otherwise it is sampled from the distribution.

julia> using TDLM.Simulate

julia> zeros(3, 3) + Noise();

julia> ones(Int16, 3, 3) + Noise([1, 2, 3]);

julia> import Distributions

julia> zeros(3, 3) + Noise(Distributions.Beta());

julia> using StableRNGs # results above are suppressed, here reproducible:

julia> zeros(2, 2) + Noise(Distributions.Normal(), StableRNG(42))
2×2 Matrix{Float64}:
 -0.670252  1.37363
  0.447122  1.30954

TDLM.Simulate.rand_cov — Function

rand_cov(k::Int, [rng = Random.GLOBAL_RNG])

Generate a random covariance matrix of size k×k.

julia> using TDLM.Simulate

julia> rand_cov(3, StableRNGs.StableRNG(42))
3×3 Matrix{Float64}:
  1.39477     0.156631   -0.0456395
  0.156631    1.54293    -0.0163959
 -0.0456395  -0.0163959   0.797937